Re <aside,
It appeared that aside tag can be used as a sidebar with some
information that is NOT related to the main content.
The element can be used for typographical effects like pull quotes or
sidebars, for advertising, for groups of nav elements, and for other
content that is considered separate from the main content of the page.
I changed <aside> to <div>